openmp with icc
Former-commit-id: a4b41a205318a4a9f2a81f4bb1d70f4ccc2c960d

+++ user/psava/Mparadd.c
@@ -175,7 +175,7 @@ static void add_float (bool   collect,
 		       float  add,
                        int    ompchunk)
 {
-    size_t jbuf;
+    int jbuf;
     float f;
 
     if (collect) {
